Update to TCL v2
and address an unexpected SRID mismatch - this isn't in SRID 4326 (WGS84) at all!  It's actually in SRID 3857, which means we need the `-s {from_srid}:{to_srid}` option to `shp2pgsql`, which means we can't use `-D`, which means this is a bit slower now as it does single-record inserts.  *sigh*